+++ Sprint1/src/operations/RotateElements.java
@@ -41,10 +41,6 @@ public class RotateElements {
 		//for each in arraylist do
 		RotateTransition rtrans = new RotateTransition(Duration.millis(250),node);   //make a new rotateTransform 1/20 of a second duration
 		rtrans.setFromAngle(0);                                                  //start angle of rotation
-		for(int i =0; i<3; i++){
-			System.out.println("test");
-			System.out.println("hrgtdcnhtbgfdwegrhtgrfv");
-		}
 		switch(facing){
 		case 0:
 			switch(turnTo){
